+++ src/atoms/ToggleButton/ToggleButton.jsx
@@ -1,4 +1,4 @@
-import { useToggleStore } from '@/hooks/useStore';
+import { useToggleStore } from '@/zustand/useStore';
 import { faToggleOn, faToggleOff } from '@fortawesome/free-solid-svg-icons';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
